Use Int64.
Former-commit-id: e397f171ad3861c8381a96ab1b3fdf96069c65c7
Former-commit-id: de701ed54d98044854fd10663c07ff091fcb974a

+++ simulation.jl
@@ -126,7 +126,7 @@ end
 
 const abstol = 1e-7
 const reltol = 1e-7
-const maxiters = Int(1e10)
+const maxiters = Int64(1e10)
 
 function simulate(m::Model, N::Int, Ts::Float64; kwargs...)
   let
